create-img: force selection of libverto-tevent
8.2: include libverto-tevent instead of libverto-libevent, and avoid
specifying explicitly packages that get pulled as dependencies

8.3: explicitly specify libverto-tevent to be sure

The only known libverto user in XCP-ng installer is gssproxy, whose
code does not care about which libverto backend will be used, so the
choice of which one to use is delegated at runtime to libverto.

libverto has a "default backend" configured, which will be used if
available, and this is libverto-tevent in this version.  Having this
package installed is the only way to ensure the same lib will be used
by gssproxy (and any other similar software, if any), regardless of
whether libverto-libevent is also installed in an install image
modified by a third-party.

This is still not perfect, as having libverto-glib installed
apparently causes libverto-tevent (for some reason) to also load
libglib.so.

Also note that CH 8.2 and current previews of XS 8 all use
libverto-libevent, we are diverging here for consistency with the
installed product.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>