create-installimg: honor --define-repo flags
The code handling this option was in code that ended up not being shared
between the two scripts, because setup of yum and yumdownloader was
different enough to cause more issues than it helped.

Split setup_yum_repos out of setup_yum_download, so create-installimg
can use it, and remove code duplication that ended up missing part of
the code.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>