8.2 template: add XenServer's GPG keys from the initial 8.2.1 ISO
It's not that we really need them, but they're listed in .treeinfo, so
they need to be present. We could have opted for removing them because
nothing is checked against them during the installation, but then the
resulting installed system would slightly differ from a system installed
with the initial 8.2.1 ISO and then updated through yum. The GPG keys
listed in .treeinfo are indeed copied to the installed system.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>